fix: 빌드 에러 수정 - 넥스트 버전에 따른 방식 변경
