+++ src/app/(with-header)/page.tsx
@@ -1,26 +1,5 @@
-'use client';
+import BasePage from '@/app/(with-header)/components/BasePage';
 
-import { useSearchParams } from 'next/navigation';
-import BannerSection from '@/app/(with-header)/components/BannerSection';
-import PopularExperiences from '@/app/(with-header)/components/PopularExperiences';
-import ExperienceList from '@/app/(with-header)/components/ExperienceList';
-
-export default function HomePage() {
-  const searchParams = useSearchParams();
-  const keyword = searchParams.get('q') ?? '';
-  const isSearchMode = Boolean(keyword.trim());
-
-  return (
-    <main>
-      <BannerSection keyword={keyword} />
-      {isSearchMode ? (
-        <ExperienceList keyword={keyword} isSearchMode />
-      ) : (
-        <>
-          <PopularExperiences />
-          <ExperienceList />
-        </>
-      )}
-    </main>
-  );
+export default function Page() {
+  return <BasePage />;
 }
